Add files via upload
Updated up to the second checkpoint. This includes updates to the Animal and Zoo classes. I have made it so that we can find the total area for each of the habitats, and it will print that information out. The next portion seems like it may be altered due to how I structured my classes.